+++ src/test/scala/edu/ie3/simona/service/results/ResultServiceProxySpec.scala
@@ -32,6 +32,14 @@ class ResultServiceProxySpec
 
   "The ResultServiceProxy" should {
 
+    val allExpected = Seq(
+      dummyNodeResultModel,
+      dummySwitchResultModel,
+      dummyLineResultModel,
+      dummyTrafo2WResultModel,
+      inputModel,
+    )
+
     "answer request for results correctly without waiting for results" in {
       val resultProvider = TestProbe[ResultResponse]("listener")
 
@@ -55,13 +63,9 @@ class ResultServiceProxySpec
       val resultProxy = spawn(ResultServiceProxy(Seq.empty, startTime, 10))
 
       // tells the proxy to wait for the results of dummyInputModel for tick 3600L
-      resultProxy ! ExpectResult(Seq(dummyInputModel), 3600L)
+      resultProxy ! ExpectResult(Seq(dummyNodeResultModel), 3600L)
 
-      resultProxy ! RequestResult(
-        Seq(dummyInputModel, inputModel),
-        3600L,
-        resultProvider.ref,
-      )
+      resultProxy ! RequestResult(allExpected, 3600L, resultProvider.ref)
 
       // still waiting for results
       resultProvider.expectNoMessage()
@@ -76,12 +80,10 @@ class ResultServiceProxySpec
 
       // no results for three winding transformers, because the proxy is not told to wait and the results was not received beforehand
       resultProvider.expectMessageType[ResultResponse].results shouldBe Map(
-        dummyInputModel -> List(
-          dummyNodeResult,
-          dummySwitchResult,
-          dummyLineResult,
-          dummyTrafo2wResult,
-        )
+        dummyNodeResultModel -> List(dummyNodeResult),
+        dummySwitchResultModel -> List(dummySwitchResult),
+        dummyLineResultModel -> List(dummyLineResult),
+        dummyTrafo2WResultModel -> List(dummyTrafo2wResult),
       )
     }
 
@@ -91,13 +93,9 @@ class ResultServiceProxySpec
       val resultProxy = spawn(ResultServiceProxy(Seq.empty, startTime, 10))
 
       // tells the proxy to wait for the results with dumyInputModel for tick 3600L
-      resultProxy ! ExpectResult(Seq(dummyInputModel), 3600L)
+      resultProxy ! ExpectResult(Seq(dummyNodeResultModel), 3600L)
 
-      resultProxy ! RequestResult(
-        Seq(dummyInputModel, inputModel),
-        3600L,
-        resultProvider.ref,
-      )
+      resultProxy ! RequestResult(allExpected, 3600L, resultProvider.ref)
 
       // receiving three winding results for port B and C beforehand
       resultProxy ! PowerFlowResultEvent(
@@ -121,12 +119,10 @@ class ResultServiceProxySpec
 
       // receives three winding result, because all partial results are present
       resultProvider.expectMessageType[ResultResponse].results shouldBe Map(
-        dummyInputModel -> List(
-          dummyNodeResult,
-          dummySwitchResult,
-          dummyLineResult,
-          dummyTrafo2wResult,
-        ),
+        dummyNodeResultModel -> List(dummyNodeResult),
+        dummySwitchResultModel -> List(dummySwitchResult),
+        dummyLineResultModel -> List(dummyLineResult),
+        dummyTrafo2WResultModel -> List(dummyTrafo2wResult),
         inputModel -> List(expected),
       )
     }
@@ -137,13 +133,19 @@ class ResultServiceProxySpec
       val resultProxy = spawn(ResultServiceProxy(Seq.empty, startTime, 10))
 
       // tells the proxy to wait for the results of dumyInputModel for tick 3600L
-      resultProxy ! ExpectResult(Seq(dummyInputModel), 3600L)
+      resultProxy ! ExpectResult(Seq(dummyNodeResultModel), 3600L)
 
       // tells the proxy to also wait for the results of inputModel for tick 3600L
       resultProxy ! ExpectResult(Seq(inputModel), 3600L)
 
       resultProxy ! RequestResult(
-        Seq(dummyInputModel, inputModel),
+        Seq(
+          dummyNodeResultModel,
+          dummySwitchResultModel,
+          dummyLineResultModel,
+          dummyTrafo2WResultModel,
+          inputModel,
+        ),
         3600L,
         resultProvider.ref,
       )
@@ -173,12 +175,10 @@ class ResultServiceProxySpec
 
       // no results for three winding transformers, because the proxy is not told to wait and the results was not received beforehand
       resultProvider.expectMessageType[ResultResponse].results shouldBe Map(
-        dummyInputModel -> List(
-          dummyNodeResult,
-          dummySwitchResult,
-          dummyLineResult,
-          dummyTrafo2wResult,
-        ),
+        dummyNodeResultModel -> List(dummyNodeResult),
+        dummySwitchResultModel -> List(dummySwitchResult),
+        dummyLineResultModel -> List(dummyLineResult),
+        dummyTrafo2WResultModel -> List(dummyTrafo2wResult),
         inputModel -> List(expected),
       )
     }
@@ -197,15 +197,47 @@ class ResultServiceProxySpec
         Seq.empty,
       )
 
-      // all results have the same uuid, therefore, all result a grouped to this uuid
+      // all results are mapped to their uuid
       listener.expectMessageType[ResultResponse].results shouldBe Map(
-        dummyInputModel -> List(
-          dummyNodeResult,
-          dummySwitchResult,
-          dummyLineResult,
-          dummyTrafo2wResult,
-        )
+        dummyNodeResultModel -> List(dummyNodeResult),
+        dummySwitchResultModel -> List(dummySwitchResult),
+        dummyLineResultModel -> List(dummyLineResult),
+        dummyTrafo2WResultModel -> List(dummyTrafo2wResult),
       )
+    }
+
+    "correctly handle unchanged grid result" in {
+      val listener = TestProbe[ResultResponse]("listener")
+
+      val resultProxy =
+        spawn(ResultServiceProxy(Seq(listener.ref), startTime, 10))
+
+      resultProxy ! PowerFlowResultEvent(
+        Seq(dummyNodeResult),
+        Seq(dummySwitchResult),
+        Seq(dummyLineResult),
+        Seq(dummyTrafo2wResult),
+        Seq.empty,
+      )
+
+      // all results are mapped to their uuid
+      listener.expectMessageType[ResultResponse].results shouldBe Map(
+        dummyNodeResultModel -> List(dummyNodeResult),
+        dummySwitchResultModel -> List(dummySwitchResult),
+        dummyLineResultModel -> List(dummyLineResult),
+        dummyTrafo2WResultModel -> List(dummyTrafo2wResult),
+      )
+
+      resultProxy ! PowerFlowResultEvent(
+        Seq(dummyNodeResult2),
+        Seq.empty,
+        Seq.empty,
+        Seq.empty,
+        Seq.empty,
+      )
+
+      // no unchanged or new results received
+      listener.expectNoMessage()
     }
 
     "correctly handle three winding transformer result events" in {